move time period <-> time functions to util/time
